Triton GEMM: promote M/N pointer offsets to int64 to fix >2^31 addressing
matmul_kernel and mxfp8_matmul_kernel compute pointer offsets as
  ptrs = base + offs_m[:, None] * stride_m + offs_k[None, :] * stride_k
Triton's default int32 arithmetic overflows silently when the product
(M-1) * stride exceeds 2^31, producing garbage pointers and — during
autotune benchmarking — HIP faults that surface as SIGABRT.

Concrete repro:
  test_float8_current_scaling_exact.py::TestFP8CurrentScalingLargeNumel::
  test_fp8_current_scaling_linear_large_numel_e4m3[bs33-fp32]
  M=143616, K=15360 -> (M-1)*K = 2,205,926,400 > 2^31 = 2,147,483,648.
Suite ran as F...F. then aborted the pytest process during the Triton
autotune of a benchmark launch. bs32 (M=139264, (M-1)*K = 2,139,095,040)
sits just under 2^31 and passes; bs33 tips over. The docstring on the
test itself flags this class of bug: "Regression for 32-bit numel
overflow when total elements > 2^31."

Fix: cast the M-dim and N-dim offset tensors to tl.int64 exactly at the
pointer-computation sites -- a_ptrs, b_ptrs, c_ptrs, and (mxfp8 kernel)
a_scale_ptrs and b_scale_ptrs. K-dim offsets stay int32 (bounded by
BLOCK_SIZE_K). Mask comparisons (offs_m < M, offs_n < N) also stay
int32 so the per-tile bounds check is not slowed.

After the fix: full test_float8_current_scaling_exact.py completes
end-to-end (12.4s, 5 passed / 2 failed -- the two remaining failures
are the separate HYBRID-recipe gate, not related to this bug).

No regression on the 4 direct Triton GEMM suites:
  test_gemm_triton.py             450 passed / 606 skipped
  test_gemm_triton_generic_fp8.py  21 passed
  test_te_generic_gemm_triton.py  210 passed /  72 skipped
  mxfp8/                            4 passed

+++ transformer_engine/pytorch/gemm_triton.py
@@ -1048,9 +1048,10 @@ def mxfp8_matmul_kernel(
     offs_bn = pid_n * BLOCK_SIZE_N + tl.arange(0, BLOCK_SIZE_N)
     offs_k = tl.arange(0, BLOCK_SIZE_K)
 
-    # Data pointers
-    a_ptrs = a_ptr + (offs_am[:, None] * stride_am + offs_k[None, :] * stride_ak)
-    b_ptrs = b_ptr + (offs_k[:, None] * stride_bk + offs_bn[None, :] * stride_bn)
+    # Data pointers. Promote M/N offsets to int64 before multiplying by
+    # potentially large strides — see matmul_kernel for the overflow rationale.
+    a_ptrs = a_ptr + (offs_am[:, None].to(tl.int64) * stride_am + offs_k[None, :] * stride_ak)
+    b_ptrs = b_ptr + (offs_k[:, None] * stride_bk + offs_bn[None, :].to(tl.int64) * stride_bn)
 
     # K-loop
     num_k_blocks = tl.cdiv(K, BLOCK_SIZE_K)
@@ -1095,7 +1096,7 @@ def mxfp8_matmul_kernel(
         k_block_start = k * (BLOCK_SIZE_K // VEC_SIZE)
         offs_a_scale_k = k_block_start + tl.arange(0, BLOCK_SIZE_K // VEC_SIZE)
 
-        a_scale_ptrs = a_scale_ptr + (offs_am[:, None] * stride_a_scale_m +
+        a_scale_ptrs = a_scale_ptr + (offs_am[:, None].to(tl.int64) * stride_a_scale_m +
                                        offs_a_scale_k[None, :] * stride_a_scale_k)
 
         # Check bounds for scale loading
@@ -1108,7 +1109,7 @@ def mxfp8_matmul_kernel(
         # B scale layout: [N, K//32] (new dot_scaled API -- "Do NOT transpose rhs_scale")
         # For tl.dot_scaled we need [BLOCK_SIZE_N, BLOCK_SIZE_K//32]
         offs_b_scale_k = k_block_start + tl.arange(0, BLOCK_SIZE_K // VEC_SIZE)
-        b_scale_ptrs = b_scale_ptr + (offs_bn[:, None] * stride_b_scale_n +
+        b_scale_ptrs = b_scale_ptr + (offs_bn[:, None].to(tl.int64) * stride_b_scale_n +
                                        offs_b_scale_k[None, :] * stride_b_scale_k)
 
         mask_b_scale_n = offs_bn < N
@@ -1140,7 +1141,8 @@ def mxfp8_matmul_kernel(
     # Store output (convert to target dtype)
     offs_cm = pid_m * BLOCK_SIZE_M + tl.arange(0, BLOCK_SIZE_M)
     offs_cn = pid_n * BLOCK_SIZE_N + tl.arange(0, BLOCK_SIZE_N)
-    c_ptrs = c_ptr + stride_cm * offs_cm[:, None] + stride_cn * offs_cn[None, :]
+    # int64 promotion (see matmul_kernel).
+    c_ptrs = c_ptr + stride_cm * offs_cm[:, None].to(tl.int64) + stride_cn * offs_cn[None, :]
     c_mask = (offs_cm[:, None] < M) & (offs_cn[None, :] < N)
 
     c = accumulator.to(c_ptr.type.element_ty)
@@ -1292,8 +1294,13 @@ def matmul_kernel(
     offs_k = tl.arange(0, BLOCK_SIZE_K)
     offs_am = (pid_m * BLOCK_SIZE_M + tl.arange(0, BLOCK_SIZE_M)) % M
     offs_bn = (pid_n * BLOCK_SIZE_N + tl.arange(0, BLOCK_SIZE_N)) % N
-    a_ptrs = a_ptr + (offs_am[:, None] * stride_am + offs_k[None, :] * stride_ak)
-    b_ptrs = b_ptr + (offs_k[:, None] * stride_bk + offs_bn[None, :] * stride_bn)
+    # Promote the M/N offsets to int64 before multiplying by a stride that may
+    # be as large as K or N: offs_am*stride_am overflows int32 whenever
+    # (M-1)*K > 2^31 (e.g. M=143616, K=15360 → 2.2G). The mask arithmetic
+    # (offs_* < M/N) stays int32 for speed; only the address computation
+    # widens.
+    a_ptrs = a_ptr + (offs_am[:, None].to(tl.int64) * stride_am + offs_k[None, :] * stride_ak)
+    b_ptrs = b_ptr + (offs_k[:, None] * stride_bk + offs_bn[None, :].to(tl.int64) * stride_bn)
 
     # -----------------------------------------------------------
     # Iterate to compute a block of the C matrix.
@@ -1360,7 +1367,8 @@ def matmul_kernel(
 
     offs_cm = pid_m * BLOCK_SIZE_M + tl.arange(0, BLOCK_SIZE_M)
     offs_cn = pid_n * BLOCK_SIZE_N + tl.arange(0, BLOCK_SIZE_N)
-    c_ptrs = c_ptr + stride_cm * offs_cm[:, None] + stride_cn * offs_cn[None, :]
+    # Same int64 promotion as the A/B pointers above: M*N can exceed 2^31.
+    c_ptrs = c_ptr + stride_cm * offs_cm[:, None].to(tl.int64) + stride_cn * offs_cn[None, :]
     c_mask = (offs_cm[:, None] < M) & (offs_cn[None, :] < N)
 
     # β accumulation: D = α·(A·B) + bias + β·C. Fold existing C in *before*
